refactor: type 정리 및 import문 정리
