+++ apps/web/src/components/auction-detail/container.tsx
@@ -8,7 +8,7 @@
  * 5. ItemImages prioty 처리
  */
 
-import { useEffect } from 'react';
+import { useEffect, useRef } from 'react';
 
 import { useParams } from 'next/navigation';
 
@@ -30,17 +30,11 @@ import { cn } from '@/lib/cn';
 import { ItemInfo } from '@/lib/types/auction';
 
 const AuctionDetailContainer = () => {
-  const userId = undefined;
-
+  const bidTableRef = useRef<HTMLDivElement>(null);
   const params = useParams();
   const { id } = params;
 
-  const {
-    data: auctionDetailData,
-    isLoading,
-    error,
-    refetch,
-  } = useAuctionDetail(id as string, userId);
+  const { data: auctionDetailData, isLoading, error, refetch } = useAuctionDetail(id as string);
 
   useEffect(() => {
     refetch();
@@ -118,13 +112,18 @@ const AuctionDetailContainer = () => {
         </ProfileCard>
         <ItemSummary item={item} id={id as string} />
         <ItemDescription item={item} />
-        <section className="space-y-2 pb-10 pt-6">
+        <section ref={bidTableRef} className="space-y-2 pb-10 pt-6">
           <h3 className="mb-2.5 text-base font-bold">입찰 현황</h3>
           <BidTable />
         </section>
       </article>
       <aside className="sticky bottom-0 z-50 w-full">
-        <BidForm auctionId={auction.id} currentPrice={item.currentPrice} endTime={item.endTime} />
+        <BidForm
+          auctionId={auction.id}
+          currentPrice={item.currentPrice}
+          endTime={item.endTime}
+          bidTableRef={bidTableRef}
+        />
       </aside>
     </>
   );
