+++ apps/web/src/lib/types/bid.ts
@@ -1,3 +1,5 @@
+import { Dispatch, SetStateAction } from 'react';
+
 interface BidType {
   bid_id: string;
   bidder_id: string;
@@ -10,4 +12,29 @@ interface BidsListType {
   bids: BidType[];
 }
 
-export type { BidType, BidsListType };
+//Bid-form props type
+interface BidBaseProps {
+  auctionId: string;
+  currentPrice: number;
+  endTime: string | Date; // 경매 종료 시간
+}
+
+interface BidFormBottomProps extends BidBaseProps {
+  isExpired: boolean;
+  isBidding: boolean;
+  onChange: Dispatch<SetStateAction<boolean>>;
+}
+
+interface BidFormProps extends BidBaseProps {
+  bidTableRef: React.RefObject<HTMLDivElement | null>;
+}
+
+interface BidInputProps {
+  currentPrice: number;
+  minUnit: number;
+  bidPrice: number | null;
+  isBidding: boolean;
+  onChange: (price: number | null) => void;
+}
+
+export type { BidType, BidsListType, BidInputProps, BidFormBottomProps, BidFormProps };
